Add support for setting MTU on vswitch ports
When defining an external connection for a vswitch in hiera, you
can define an MTU setting for the interface. For a single interface
you could add the following in hiera:

profile::baseconfig::network::bridges:
  bridgename:
    external:
      type: 'interface'
      name: 'eno1'
      mtu: 9000
      driver: 'bnx2x'

For a bond the similar setup would look like so in hiera:

profile::baseconfig::network::bridges:
  bridgename:
    external:
      type: 'bond'
      mtu: 9000
      members:
        eno1:
          driver: 'bnx2x'
        eno2:
          driver: 'bnx2x'

NOTE: Du to a current bug in netplan we need to match a specific interface to
set the MTU-value. When using vswitches as we do, we need to match both
mac-address and drivername to sit with a single interface. This is the reason
behind the driver-option.