Add possibility to add static-routes for bird.
This is useful if bird should peer with adresses on networks which are not
directly connected.

New hiera-key:
`profile::bird::ipv4::static` - Hash where the keys are a network CIDR, and the
value is a new hash with at least one item: next-hop.

This might look like so in hiera:
profile::bird::ipv4::static:
  10.37.0.140/32:
    next-hop: '10.0.28.2'
  10.37.0.141/32:
    next-hop: '10.0.28.3'